Fixed for watchdog, updated di-remove

+++ scripts/servers_di_remove.sh
@@ -1,23 +1,33 @@
 #source /dbagiga/setenv.sh
 
-systemctl stop odsxkafka.service
-sleep 2
-systemctl stop odsxzookeeper.service
-sleep 2
-systemctl stop telegraf
-yum -y remove telegraf
-sleep 2
+wantToRemoveKafka=$1
+wantToRemoveZk=$2
+wantToRemoveTelegraf=$3
 
-rm -rf $KAFKA_LOGS_PATH
-rm -rf $KAFKA_DATA_PATH
-rm -rf $KAFKAPATH
+if [ "$wantToRemoveKafka" == "y" ]; then
+  systemctl stop odsxkafka.service
+  sleep 2
+  rm -rf $KAFKA_LOGS_PATH
+  rm -rf $KAFKA_DATA_PATH
+  rm -rf $KAFKAPATH
+  rm -rf install install.tar /dbagiga/setenv.sh /usr/local/bin/st*_kafka.sh /etc/systemd/system/kafka.service /etc/systemd/system/odsxkafka.service
+fi
 
-rm -rf $ZOOKEEPER_DATA_PATH
-rm -rf $ZOOKEEPER_LOGS_PATH
-rm -rf $ZOOKEEPERPATH
+if [ "$wantToRemoveZk" == "y" ]; then
+  systemctl stop odsxzookeeper.service
+  sleep 2
+  rm -rf $ZOOKEEPER_DATA_PATH
+  rm -rf $ZOOKEEPER_LOGS_PATH
+  rm -rf $ZOOKEEPERPATH
+  rm -rf /usr/local/bin/st*_zookeeper.sh /etc/systemd/system/odsxzookeeper.service
+fi
+if [ "$wantToRemoveTelegraf" == "y" ]; then
+  systemctl stop telegraf
+  yum -y remove telegraf
+  sleep 2
+fi
 
-rm -rf install install.tar /dbagiga/setenv.sh /usr/local/bin/st*_kafka.sh /etc/systemd/system/kafka.service
-rm -rf /usr/local/bin/st*_zookeeper.sh /etc/systemd/system/odsxkafka.service /etc/systemd/system/odsxzookeeper.service
+
 
 #DIM - Services
 systemctl stop di-mdm.service
